Handle new-style m-lab name format lookups.
The blipdns.apenwarr.ca nameserver no longer gets a full list of m-lab
sites, since they started filtering the names in their API. This made
it so a lot of the DNS lookups would fail, leading to us not actually
trying an http request after looking up a host.

Change the DNS name format so that we rely less on m-lab's precise set
of hostnames.